fix files according to arduino library specification
