+++ README.md
@@ -1,22 +1,3 @@
-TroykaIMU
-==========
+# TroykaIMU
 
-Библиотека для Arduino, позволяющая управлять [IMU-сенсором на 10 степеней свободы (Troyka-модуль)](http://amperka.ru/product/troyka-imu-10-dof)
-от [Амперки](http://amperka.ru/).
-
-Установка библиотеки
-====================
-
-В Arduino IDE выберите пункт меню «Скетч» → «Импортировать библиотеку» →
-«Добавить библиотеку…». В появившемся окне выберите скачаный архив с
-библиотекой. Установка завершена.
-
-Подробности и примеры работы для [IMU-сенсора на 10 степеней свободы (Troyka-модуль)](http://wiki.amperka.ru/продукты:troyka-imu-10-dof).
-
-Подробности и примеры работы для [акселерометра (Troyka-модуль)](http://wiki.amperka.ru/продукты:troyka-accelerometer).
-
-Подробности и примеры работы для [гироскопа (Troyka-модуль)](http://wiki.amperka.ru/продукты:troyka-gyro).
-
-Подробности и примеры работы для [компаса (Troyka-модуль)](http://wiki.amperka.ru/продукты:troyka-compass).
-
-Подробности и примеры работы для [барометра (Troyka-модуль)](http://wiki.amperka.ru/продукты:troyka-barometer).
+Arduino library to interface with the [Amperka IMU 10-DOF Sensor](https://amperka.ru/product/troyka-imu-10-dof).
