+++ library.properties
@@ -1,0 +1,9 @@
+name=Troyka-IMU
+version=1.1.0
+author=Igor Dementiev <igor@amperka.ru>
+maintainer=Amperka <dev@amperka.com>
+sentence=Arduino library for Amperka IMU-sensor.
+paragraph=Allows you to read the accelerometer, magnetometer, gyroscope and barometer values from the Amperka IMU 10-DOF Sensor.
+category=Sensors
+url=https://github.com/amperka/Troyka-IMU
+architectures=*
