+++ src/stmhw.cpp
