+++ src/stmhw.h
